Initial commit.
Added plugin discovery for @ContentEntityTypeBundle annotations. These
allow for content entity type bundles to be instantiated with
subclasses of the base entity class.

Added an SQL entity storage driver for extendable content entities
which can be used as a storage handler on entities that allow for
subclasses for each bundle.